# openapi/access-v1.yaml
openapi: 3.1.0
info:
  title: Prison Policy Access API
  version: 1.0.0
  description: Versioned API for the managed Microsoft Access client.
...
security:
  - AccessBearer: []
paths:
...
  /api/browser-handoffs/redeem:
    post:
      operationId: redeemReviewLabHandoff
      tags: [Browser Handoff]
      security: []
      parameters:
        - { $ref: "#/components/parameters/XRequestId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BrowserHandoffRedeemRequest" }
      responses:
        "200":
          description: Handoff redeemed into an HttpOnly browser session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BrowserHandoffRedeemSuccessEnvelope" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/HandoffInvalid" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/DependencyUnavailable" }
components:
  securitySchemes:
    AccessBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
...
    XRequestId:
      name: X-Request-ID
      in: header
      required: false
      schema:
        $ref: "#/components/schemas/RequestId"
...
  headers:
    XRequestId:
      description: Validated request correlation identifier.
      schema:
        $ref: "#/components/schemas/RequestId"
  responses:
    ValidationFailed:
      description: Closed request validation failed
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    RateLimited:
      description: Authentication rate limit exceeded
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestId"
        Retry-After:
          description: Seconds until every exceeded login dimension is safe to retry.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    DependencyUnavailable:
      description: Authentication persistence dependency is unavailable
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
    HandoffInvalid:
      description: The one-time handoff is invalid, expired, or already redeemed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    RequestId:
      type: string
      pattern: '^[A-Za-z0-9_-]{8,64}$'
      example: request_20260812_000001
    UtcTimestamp:
      type: string
      format: date-time
      example: '2026-08-12T18:30:00Z'
...
    GenericSuccessEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id, server_time, api_version]
      properties:
        data:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
        request_id:
          $ref: "#/components/schemas/RequestId"
        server_time:
          $ref: "#/components/schemas/UtcTimestamp"
        api_version:
          type: string
          const: v1
...
    BrowserHandoffRedeemRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 40
          maxLength: 512
          writeOnly: true
          example: example-token-not-valid
    BrowserHandoffRedeemData:
      type: object
      additionalProperties: false
      required: [redeemed]
      properties:
        redeemed: { type: boolean, const: true }
...
    BrowserHandoffRedeemSuccessEnvelope:
      allOf: [{ $ref: "#/components/schemas/GenericSuccessEnvelope" }, { type: object, properties: { data: { $ref: "#/components/schemas/BrowserHandoffRedeemData" } } }]
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, request_id, server_time, api_version]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        request_id:
          $ref: "#/components/schemas/RequestId"
        server_time:
          $ref: "#/components/schemas/UtcTimestamp"
        api_version:
          type: string
          const: v1